requirements.txt
allure-pytest==2.13.5
allure-python-commons==2.13.5
attrs==25.1.0
bcrypt==5.0.0
cachetools==6.2.5
certifi==2025.6.15
cffi==2.0.0
charset-normalizer==3.4.1
cryptography==46.0.3
cssselect==1.3.0
cssutils==2.11.1
Faker==40.1.2
h11==0.14.0
idna==3.10
iniconfig==2.0.0
invoke==2.2.1
jsonpath==0.82.2
loguru==0.7.3
lxml==6.0.2
more-itertools==10.8.0
MouseInfo==0.1.3
outcome==1.3.0.post0
packaging==24.2
paramiko==4.0.0
pluggy==1.5.0
premailer==3.10.0
PyAutoGUI==0.9.54
pycparser==3.0
PyGetWindow==0.0.9
PyMsgBox==2.0.1
PyMySQL==1.1.2
PyNaCl==1.6.2
pyperclip==1.11.0
PyRect==0.2.0
PyScreeze==1.0.1
PySocks==1.7.1
pytest==8.3.5
pytest-rerunfailures==16.1
pytweening==1.2.0
PyYAML==6.0.2
requests==2.32.3
requests-toolbelt==1.0.0
schedule==1.2.2
selenium==4.29.0
sniffio==1.3.1
sortedcontainers==2.4.0
sshtunnel==0.4.0
trio==0.29.0
trio-websocket==0.12.2
typing_extensions==4.12.2
urllib3==2.3.0
websocket-client==1.8.0
wsproto==1.2.0
yagmail==0.15.293
